Moved NavigationRoutes to utils
