+++ src/ros/nova-gui/nova-gui/src/utils/NavigationRoutes.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from "react";
 import { Aperture, Camera, Circle, Globe, Home, Image, List, Map, Terminal, Tv } from "react-feather";
-import { URCCompModes } from "../../../views/shared/CamerasPage/CameraPageConstants";
+import { URCCompModes } from "../views/shared/CamerasPage/CameraPageConstants";
 
 interface NavigationEntry {
   title: string;
